requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
shap>=0.44.0
matplotlib>=3.7.0
streamlit>=1.34.0
pyyaml>=6.0
pytest>=7.4.0